Move additional types to primitive & rpc
