+++ packages/api-derive/src/accounts/indexes.ts
@@ -5,7 +5,7 @@
 import { Observable } from 'rxjs';
 import { map, switchMap } from 'rxjs/operators';
 import { ApiInterface$Rx } from '@polkadot/api/types';
-import { ENUMSET_SIZE } from '@polkadot/types/type/AccountIndex';
+import { ENUMSET_SIZE } from '@polkadot/types/primitive/AccountIndex';
 import { AccountId, AccountIndex } from '@polkadot/types';
 
 import { drr } from '../util/drr';
