+++ packages/types/src/primitive/AccountId.spec.ts
@@ -7,7 +7,7 @@ import U8a from '../codec/U8a';
 import Vector from '../codec/Vector';
 import jsonVec from '../json/AccountIdVec.001.json';
 import AccountId from './AccountId';
-import StorageData from '../primitive/StorageData';
+import StorageData from './StorageData';
 
 describe('AccountId', () => {
   describe('defaults', () => {
