+++ packages/types/src/primitive/Origin.ts
@@ -2,7 +2,7 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import Null from '../primitive/Null';
+import Null from './Null';
 
 /**
  * @name Origin
