+++ packages/types/src/type/AttestedCandidate.ts
@@ -8,11 +8,11 @@ import EnumType from '../codec/EnumType';
 import Struct from '../codec/Struct';
 import Tuple from '../codec/Tuple';
 import Vector from '../codec/Vector';
+import AccountId from '../primitive/AccountId';
 import Bytes from '../primitive/Bytes';
+import Hash from '../primitive/Hash';
 import Null from '../primitive/Null';
 import U64 from '../primitive/U64';
-import AccountId from './AccountId';
-import Hash from './Hash';
 import ParaId from './ParaId';
 import Signature from './Signature';
 import SessionKey from './SessionKey';
