+++ packages/types/src/type/Bft.ts
@@ -5,8 +5,8 @@
 import { AnyU8a } from '../types';
 
 import Tuple from '../codec/Tuple';
+import Hash from '../primitive/Hash';
 import AuthorityId from './AuthorityId';
-import Hash from './Hash';
 import Signature from './Signature';
 
 export type BftAuthoritySignatureValue = [AnyU8a, AnyU8a];
