+++ packages/types/src/type/CodeHash.ts
@@ -2,7 +2,7 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import Hash from './Hash';
+import Hash from '../primitive/Hash';
 
 /**
  * @name CodeHash
