--- packages/types/src/type/EventIndex.ts
@@ -1,13 +1,0 @@
-// Copyright 2017-2019 @polkadot/types authors & contributors
-// This software may be modified and distributed under the terms
-// of the Apache-2.0 license. See the LICENSE file for details.
-
-import U32 from '../primitive/U32';
-
-/**
- * @name EventIndex
- * @description
- * The Substrate EventIndex representation as a [[U32]].
- */
-export default class EventIndex extends U32 {
-}
