+++ packages/types/src/type/Extrinsic.ts
@@ -12,9 +12,9 @@ import Compact from '../codec/Compact';
 import Struct from '../codec/Struct';
 import { FunctionMetadata } from '../Metadata/v0/Modules';
 import Method from '../primitive/Method';
-import Address from './Address';
+import Address from '../primitive/Address';
+import Hash from '../primitive/Hash';
 import ExtrinsicSignature from './ExtrinsicSignature';
-import Hash from './Hash';
 
 type ExtrinsicValue = {
   method?: Method
