+++ packages/types/src/type/StakingLedger.ts
@@ -5,7 +5,7 @@
 import Compact from '../codec/Compact';
 import Struct from '../codec/Struct';
 import Vector from '../codec/Vector';
-import AccountId from './AccountId';
+import AccountId from '../primitive/AccountId';
 import Balance from './Balance';
 import BlockNumber from './BlockNumber';
 import UnlockChunk from './UnlockChunk';
