+++ packages/types/src/type/TreasuryProposal.ts
@@ -3,7 +3,7 @@
 // of the Apache-2.0 license. See the LICENSE file for details.
 
 import Struct from '../codec/Struct';
-import AccountId from './AccountId';
+import AccountId from '../primitive/AccountId';
 import Balance from './Balance';
 
 /**
